+++ packages/core/src/components/Button/types.ts
@@ -1,0 +1,16 @@
+export enum ButtonMode {
+  CONTAINED = 'contained',
+  OUTLINED = 'outlined',
+  MINIMAL = 'minimal',
+}
+
+export enum ButtonColor {
+  PRIMARY = 'primary',
+  DANGER = 'danger',
+}
+
+export enum ButtonSize {
+  DEFAULT = 'default',
+  LARGE = 'large',
+  SMALL = 'small',
+}
