+++ packages/core/src/index.ts
@@ -3,5 +3,3 @@
  * */
 export { default as Pagination } from './components/Pagination';
 export { default as Button } from './components/Button';
-
-export default {};
