+++ packages/documents/src/Button.mdx
@@ -12,9 +12,11 @@ import { Button } from "@acme-ui/core";
 
 <Props of={Button} />
 
+<Props of={Button.Group} />
+
 ## 主要按钮
 
-###### (重点突出)：用于主行动点，一个操作区域只能有一个主按钮。
+(重点突出)：用于主行动点，一个操作区域只能有一个主按钮。
 
 <Playground>
   <div
@@ -28,7 +30,7 @@ import { Button } from "@acme-ui/core";
 
 ## 次要按钮
 
-###### (中等强调)：用于没有主次之分的一组行动点。
+(中等强调)：用于没有主次之分的一组行动点。
 
 <Playground>
   <div
@@ -50,7 +52,7 @@ import { Button } from "@acme-ui/core";
 
 ## 文字按钮
 
-###### (低强调)：文本按钮通常用于次要的操作。
+(低强调)：文本按钮通常用于次要的操作。
 
 <Playground>
   <div
@@ -68,7 +70,7 @@ import { Button } from "@acme-ui/core";
 
 ## 危险按钮
 
-###### (红色警告)：删除/移动/修改权限等危险操作，一般需要二次确认。
+(红色警告)：删除/移动/修改权限等危险操作，一般需要二次确认。
 
 <Playground>
   <div
@@ -86,7 +88,7 @@ import { Button } from "@acme-ui/core";
 
 ## 幽灵按钮
 
-###### 将其他按钮的内容反色，背景变为透明，常用在有色背景上。
+将其他按钮的内容反色，背景变为透明，常用在有色背景上。
 
 <Playground>
   <div
@@ -111,12 +113,12 @@ import { Button } from "@acme-ui/core";
 
 ## 按钮组
 
-###### 常用于多项类似操作。
+常用于多项类似操作。
 
 <Playground>
   <div style={{display: 'flex', justifyContent: "center"}}>
-    <Button.Group style={{marginRight: '80px'}}>
-      <Button mode="outlined">
+    <Button.Group style={{marginRight: '80px'}} mode="minimal">
+      <Button mode="outlined" size="large">
         1
       </Button>
       <Button mode="outlined">
@@ -141,7 +143,7 @@ import { Button } from "@acme-ui/core";
 
 ## Block 按钮
 
-###### 将按钮宽度调整为其父宽度的选项。
+将按钮宽度调整为其父宽度的选项。
 
 <Playground>
   <div>
@@ -162,7 +164,7 @@ import { Button } from "@acme-ui/core";
 
 ## 按钮尺寸
 
-###### 按钮有大、标准、小三种尺寸。
+按钮有大、标准、小三种尺寸。
 
 <Playground>
   <div
@@ -177,22 +179,4 @@ import { Button } from "@acme-ui/core";
     <Button>标准尺寸</Button>
     <Button size="small">小尺寸</Button>
   </div>
-</Playground>
-
-## 复杂一点，带有状态管理的使用
-
-<Playground>
-  {() => {
-    const [status, setStatus] = React.useState("primary");
-    function onButtonClick() {
-      setStatus(status === "primary" ? "danger" : "primary");
-    }
-    return (
-      <div>
-        <Button onClick={onButtonClick} color={status}>
-          change Status
-        </Button>
-      </div>
-    );
-  }}
 </Playground>
